feat: image and icon hosting via IPFS
The image and icon extractors now work using the InterPlanetary
File System (IPFS) instead of using imgur. This is because of
rate limit issues and possible abuse when uploading the same
image multiple times.

In comparison, IPFS addresses content by their hash, so we don't
have to keep track of if we've already uploaded an image or not.

To grab an IPFS image, an address can be inserted into a public
IPFS gateway, or a local IPFS daemon. Since we don't expect the
average person to have an IPFS daemon, in practice IPFS just acts
as a fancy caching system.

As such, I've set up a private IPFS pinning server at
[ipfs dot rhythm dot cafe] for use with orchard. Three environment variables
are needed to configure vitals to work with it.