Add new JSON Schema tests
I would like to be able to pass data straight from vitals into
postgres via postgREST without having to do data massaging and
validating beforehand in orchard. Ideally we just chuck the output
of vitals straight into a POST HTTP request. To do this, we need
to be able to trust that vitals gives us data compatible with the
database schema.

This commit adds a JSON Schema definition to the test cases
which should exactly match the fields of the Level database schema.
Actually, the test doesn't pass yet, but it should be just a bunch
of renaming.