don't fail if icon image missing
so there's a couple rdzips where they've said the icon is "icon.png",
but then "icon.png" isn't actually in the rdzip???

this commit fixes those. but it's possible that happens again for the
preview image, which is more of an issue, because I assume the preview
image is always there (since it's a required field in the publish GUI).