sha256 is now base58 encoded
This is just to reduce the size of the hash. I want to use the hash
in the URL for linking to specific levels, so reducing the length
helps with that.

i'm working on a extractor to get the "playable length" of the level,
like in chorus. not ready yet though